Split negotiation parts into focused modules
