+++ crates/transport/src/negotiation/parts/mod.rs
@@ -1,0 +1,5 @@
+mod stream_parts;
+mod try_map_error;
+
+pub use stream_parts::NegotiatedStreamParts;
+pub use try_map_error::TryMapInnerError;
